[C] Refactor detail index route redirects
When the active route has a `redirect` property, the application should
redirect the user to that URL. This commit refactors existing pages so
that this redirect can happen in one page, in `_app`, rather than in
every individual page.
