+++ pages/communities/[slug]/index.tsx
@@ -1,24 +1,9 @@
 import React from "react";
-import CommunityLayout from "components/composed/community/CommunityLayout";
-import { useRouter } from "next/router";
-import { RouteHelper } from "routes";
 import { Page } from "types/page";
 
+// This is a redirect route. See routes/baseRoutes.ts.
 const CommunityDetail: Page = () => {
-  const router = useRouter();
-  const activeRoute = RouteHelper.activeRoute();
-
-  // Nothing to see here, move on to collections
-  router.replace({
-    pathname: activeRoute?.redirect,
-    query: { ...router.query },
-  });
-
   return null;
-};
-
-CommunityDetail.getLayout = (page) => {
-  return <CommunityLayout>{page}</CommunityLayout>;
 };
 
 export default CommunityDetail;
