+++ pages/contributors/[slug]/index.tsx
@@ -1,17 +1,7 @@
-import { useRouter } from "next/router";
-import { RouteHelper } from "routes";
 import { Page } from "types/page";
 
+// This is a redirect route. See routes/baseRoutes.ts.
 const Contributor: Page = () => {
-  const router = useRouter();
-  const activeRoute = RouteHelper.activeRoute();
-
-  // Nothing to see here, moving on...
-  router.replace({
-    pathname: activeRoute?.redirect,
-    query: { ...router.query },
-  });
-
   return null;
 };
